Module redbpf::build

Building ELF

To make buildling ELF files out of eBPF code easier, and as part of the Rust build process, this module provides a number of helpers.

Currently, the headers module only know how to find kernel sources on Arch Linux, or through the KERNEL_SOURCE environment variable.

On top of building ELF files, you can also generate Rust bindings for data structures defined in C. Only structures that are defined as struct _data_[^{}]* are picked up by bindgen. This naming convention might change in the future, but has been flexible enough.

Because the compile + bindgen steps are fairly costly, they will slow down builds during development. The BuildCache struct provides a low-friction interface to only rebuild when the source files actually changed. Note that at the moment BuildCache only considers individual files, and not an entire BPF workspace. Alternative cache strategies should be easy to integrate.

A full working example of the build process might look like this:

use redbpf::build::{build, generate_bindings, cache::BuildCache, headers::headers};

fn main() -> Result<(), Error> {
    let out_dir = PathBuf::from(env::var("OUT_DIR")?);

    let headers = headers().unwrap();
    let flags = {
        let mut cflags: Vec<OsString> = redbpf::build::BUILD_FLAGS
            .iter()
            .map(OsString::from)
            .collect();

        cflags.append(&mut headers.clone());
        cflags
    };
    let bindgen_flags: Vec<String> = flags
        .iter()
        .cloned()
        .map(|f| f.into_string().unwrap())
        .collect();

    let mut cache = BuildCache::new(&out_dir);

    for file in source_files("./bpf", "c")? {
        if cache.file_changed(&file) {
            build(&flags[..], &out_dir, &file).expect("Failed building BPF plugin!");
        }
    }
    for file in source_files("./bpf", "h")? {
        if cache.file_changed(&file) {
            generate_bindings(&bindgen_flags[..], &out_dir, &file)
                .expect("Failed generating data bindings!");
        }
    }

    cache.save();
    Ok(())
}

Modules

cache

headers

Enums

Error

Constants

BUILD_FLAGS

Functions

build

generate_bindings